共通化のためのファイル分割
