+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "deploy": "sls deploy",
-    "lint": "eslint index.js",
+    "lint": "eslint *.js",
     "local": "sls offline start"
   },
   "repository": {
